+++ src/PlayerEngagement.Host/Contracts/Policies/UpdateSegmentOverridesRequest.cs
@@ -6,5 +6,5 @@ namespace PlayerEngagement.Host.Contracts.Policies;
 /// Payload for replacing the set of segment overrides under a policy key.
 /// </summary>
 public sealed record UpdateSegmentOverridesRequest {
-    public Dictionary<string, int> Overrides { get; init; } = new();
+    public Dictionary<string, long> Overrides { get; init; } = new();
 }
