+++ src/PlayerEngagement.Infrastructure/Persistence/DTOs/XpPolicyDTOs/ActivePolicyDTO.cs
@@ -10,7 +10,7 @@ public sealed record ActivePolicyDTO(
     string PolicyKey,
     string DisplayName,
     string Description,
-    int PolicyVersion,
+    long PolicyVersion,
     string Status,
     int BaseXpAmount,
     string Currency,
